% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronto.R
\docType{package}
\name{pronto}
\alias{pronto}
\alias{pronto-package}
\title{Retrieve Data from Seattle's Pronto Cycle Sharing System}
\description{
\code{pronto} is a lightweight package that retrieves up-to-date station
information for the \href{http://www.prontocycleshare.com}{Pronto Cycle
Sharing} system. Because the API simply dumps all data for the system,
\code{\link{pronto_stations}} provides all of the main functionality of this
package.
}
\seealso{
Pronto data stream: \url{https://secure.prontocycleshare.com/data/stations.json}

Data format description: \url{http://www.prontocycleshare.com/assets/pdf/JSON.pdf}
}

