% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronto_GET.R
\name{pronto_GET}
\alias{pronto_GET}
\title{GET information using the Pronto API}
\usage{
pronto_GET()
}
\value{
A list containing the response to the API call
}
\description{
\code{pronto_GET} issues queries to the Pronto API. At the moment, only
station data can be retrieved.
}
\examples{
raw <- pronto_GET()

}
\seealso{
Data stream: \url{https://secure.prontocycleshare.com/data/stations.json}
}

