% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronto_ok.R
\name{pronto_ok}
\alias{pronto_ok}
\title{Check whether a call to the Pronto API was successful}
\usage{
pronto_ok(req)
}
\arguments{
\item{req}{The results of a \code{\link{pronto_GET}} command}
}
\value{
\code{TRUE} when the request succeeded and \code{FALSE} otherwise.
}
\description{
\code{pronto_ok} checks whether a call to the pronto API was successful
}
\examples{
r <- pronto_GET()
pronto_ok(r)

}

