% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pronto_stations.R
\name{pronto_stations}
\alias{pronto_stations}
\title{Get current information for Pronto stations
\code{pronto_stations} returns current information for Pronto Cycle Share
stations. The result is a list containing the following fields:}
\usage{
pronto_stations()
}
\value{
A list containing information about the state of all Pronto stations
}
\description{
\itemize{
    \item \code{timestamp} Station information upload timestamp (ms)
    \item \code{stations} A list of information for every station in the system
    \item \code{schemeSuspended} Whether or not operations are suspended at all stations
}
}
\examples{
station_info <- pronto_stations()

}
\seealso{
Data format description: \url{http://www.prontocycleshare.com/assets/pdf/JSON.pdf}
}

